Added individual pixel modes for several random type effects

+++ src/Segment_Stuff/PixelInfoPS.h
@@ -6,8 +6,8 @@
 struct pixelInfoPS {
   CRGB color; //the color the pixel should be based on the colorMode
   uint16_t pixelLoc; //it's physical address on the strip (not local to segment set)
-  uint8_t segNum; //the number of the segment it's located in
-  uint8_t lineNum; //the line number it's on based on the maximum segment length in the set
+  uint16_t segNum; //the number of the segment it's located in
+  uint16_t lineNum; //the line number it's on based on the maximum segment length in the set
 } ;
 
 #endif
